swiss-knife-v2/requirements.txt

dnspython
requests
beautifulsoup4
